LayoutTests/fast/css/parsing-color-function-expected.txt

Test the parsing of color() function colors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing color(srgb ...)
PASS computedStyle("background-color", "color(srgb 0% 0% 0%)") is "color(srgb 0 0 0)"
PASS computedStyle("background-color", "color(srgb 10% 10% 10%)") is "color(srgb 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(srgb .2 .2 25%)") is "color(srgb 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(srgb 0 0 0 / 1)") is "color(srgb 0 0 0)"
PASS computedStyle("background-color", "color(srgb 0% 0 0 / 0.5)") is "color(srgb 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(srgb 20% 0 10/0.5)") is "color(srgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(srgb 20% 0 10/50%)") is "color(srgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(srgb 400% 0 10/50%)") is "color(srgb 1 0 1 / 0.5)"
PASS computedStyle("background-color", "color(srgb 50% -160 160)") is "color(srgb 0.5 0 1)"
PASS computedStyle("background-color", "color(srgb 50% -200 200)") is "color(srgb 0.5 0 1)"
PASS computedStyle("background-color", "color(srgb 0 0 0 / -10%)") is "color(srgb 0 0 0 / 0)"
PASS computedStyle("background-color", "color(srgb 0 0 0 / 110%)") is "color(srgb 0 0 0)"
PASS computedStyle("background-color", "color(srgb 0 0 0 / 300%)") is "color(srgb 0 0 0)"
PASS computedStyle("background-color", "color(srgb 50% -200)") is "color(srgb 0.5 0 0)"
PASS computedStyle("background-color", "color(srgb 50%)") is "color(srgb 0.5 0 0)"
PASS computedStyle("background-color", "color(srgb)") is "color(srgb 0 0 0)"
PASS computedStyle("background-color", "color(srgb 50% -200 / 0.5)") is "color(srgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(srgb 50% / 0.5)") is "color(srgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(srgb / 0.5)") is "color(srgb 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(srgb 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(srgb 40% 0 0deg)") is "rgba(0, 0, 0, 0)"

Testing color(a98-rgb ...)
PASS computedStyle("background-color", "color(a98-rgb 0% 0% 0%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 10% 10% 10%)") is "color(a98-rgb 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(a98-rgb .2 .2 25%)") is "color(a98-rgb 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 1)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 / 0.5)") is "color(a98-rgb 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 20% 0 10/0.5)") is "color(a98-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 20% 0 10/50%)") is "color(a98-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 400% 0 10/50%)") is "color(a98-rgb 1 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 50% -160 160)") is "color(a98-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200 200)") is "color(a98-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / -10%)") is "color(a98-rgb 0 0 0 / 0)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 110%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 300%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200)") is "color(a98-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50%)") is "color(a98-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(a98-rgb)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200 / 0.5)") is "color(a98-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 50% / 0.5)") is "color(a98-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb / 0.5)") is "color(a98-rgb 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 40% 0 0deg)") is "rgba(0, 0, 0, 0)"

Testing color(rec2020 ...)
PASS computedStyle("background-color", "color(rec2020 0% 0% 0%)") is "color(rec2020 0 0 0)"
PASS computedStyle("background-color", "color(rec2020 10% 10% 10%)") is "color(rec2020 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(rec2020 .2 .2 25%)") is "color(rec2020 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(rec2020 0 0 0 / 1)") is "color(rec2020 0 0 0)"
PASS computedStyle("background-color", "color(rec2020 0% 0 0 / 0.5)") is "color(rec2020 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 20% 0 10/0.5)") is "color(rec2020 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 20% 0 10/50%)") is "color(rec2020 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 400% 0 10/50%)") is "color(rec2020 1 0 1 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 50% -160 160)") is "color(rec2020 0.5 0 1)"
PASS computedStyle("background-color", "color(rec2020 50% -200 200)") is "color(rec2020 0.5 0 1)"
PASS computedStyle("background-color", "color(rec2020 0 0 0 / -10%)") is "color(rec2020 0 0 0 / 0)"
PASS computedStyle("background-color", "color(rec2020 0 0 0 / 110%)") is "color(rec2020 0 0 0)"
PASS computedStyle("background-color", "color(rec2020 0 0 0 / 300%)") is "color(rec2020 0 0 0)"
PASS computedStyle("background-color", "color(rec2020 50% -200)") is "color(rec2020 0.5 0 0)"
PASS computedStyle("background-color", "color(rec2020 50%)") is "color(rec2020 0.5 0 0)"
PASS computedStyle("background-color", "color(rec2020)") is "color(rec2020 0 0 0)"
PASS computedStyle("background-color", "color(rec2020 50% -200 / 0.5)") is "color(rec2020 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 50% / 0.5)") is "color(rec2020 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(rec2020 / 0.5)") is "color(rec2020 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(rec2020 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(rec2020 40% 0 0deg)") is "rgba(0, 0, 0, 0)"

Testing color(prophoto-rgb ...)
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0% 0%)") is "color(prophoto-rgb 0 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 10% 10% 10%)") is "color(prophoto-rgb 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(prophoto-rgb .2 .2 25%)") is "color(prophoto-rgb 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(prophoto-rgb 0 0 0 / 1)") is "color(prophoto-rgb 0 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0 0 / 0.5)") is "color(prophoto-rgb 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb 20% 0 10/0.5)") is "color(prophoto-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb 20% 0 10/50%)") is "color(prophoto-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb 400% 0 10/50%)") is "color(prophoto-rgb 1 0 1 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb 50% -160 160)") is "color(prophoto-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(prophoto-rgb 50% -200 200)") is "color(prophoto-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(prophoto-rgb 0 0 0 / -10%)") is "color(prophoto-rgb 0 0 0 / 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0 0 0 / 110%)") is "color(prophoto-rgb 0 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0 0 0 / 300%)") is "color(prophoto-rgb 0 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 50% -200)") is "color(prophoto-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 50%)") is "color(prophoto-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb)") is "color(prophoto-rgb 0 0 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 50% -200 / 0.5)") is "color(prophoto-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb 50% / 0.5)") is "color(prophoto-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(prophoto-rgb / 0.5)") is "color(prophoto-rgb 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(prophoto-rgb 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(prophoto-rgb 40% 0 0deg)") is "rgba(0, 0, 0, 0)"

Testing color(xyz ...)
PASS computedStyle("background-color", "color(xyz 0 0 0)") is "color(xyz 0 0 0)"
PASS computedStyle("background-color", "color(xyz 0 0 0 / 1)") is "color(xyz 0 0 0)"
PASS computedStyle("background-color", "color(xyz 1 1 1)") is "color(xyz 1 1 1)"
PASS computedStyle("background-color", "color(xyz 1 1 1 / 1)") is "color(xyz 1 1 1)"
PASS computedStyle("background-color", "color(xyz -1 -1 -1)") is "color(xyz -1 -1 -1)"
PASS computedStyle("background-color", "color(xyz 0.1 0.1 0.1)") is "color(xyz 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(xyz 10 10 10)") is "color(xyz 10 10 10)"
PASS computedStyle("background-color", "color(xyz .2 .2 .25)") is "color(xyz 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(xyz 0 0 0 / 0.5)") is "color(xyz 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(xyz .20 0 10/0.5)") is "color(xyz 0.2 0 10 / 0.5)"
PASS computedStyle("background-color", "color(xyz .20 0 10/50%)") is "color(xyz 0.2 0 10 / 0.5)"
PASS computedStyle("background-color", "color(xyz 0 0 0 / -10%)") is "color(xyz 0 0 0 / 0)"
PASS computedStyle("background-color", "color(xyz 0 0 0 / 110%)") is "color(xyz 0 0 0)"
PASS computedStyle("background-color", "color(xyz 0 0 0 / 300%)") is "color(xyz 0 0 0)"
PASS computedStyle("background-color", "color(xyz 1 1)") is "color(xyz 1 1 0)"
PASS computedStyle("background-color", "color(xyz 1)") is "color(xyz 1 0 0)"
PASS computedStyle("background-color", "color(xyz)") is "color(xyz 0 0 0)"
PASS computedStyle("background-color", "color(xyz 1 1 / .5)") is "color(xyz 1 1 0 / 0.5)"
PASS computedStyle("background-color", "color(xyz 1 / 0.5)") is "color(xyz 1 0 0 / 0.5)"
PASS computedStyle("background-color", "color(xyz / 50%)") is "color(xyz 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(xyz 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0% 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(xyz 40% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
